Refactor duplicate facts validation to extract the enumeration and type definitions to a separate module.
These are used as part of the initial loading of CntlrCmdLine, CntlrWinMain, api.Session etc.

Extracting these means a more minimal import can be done, improving start-up time by about 300ms.
