+++ arelle/CntlrCmdLine.py
@@ -40,7 +40,7 @@ from arelle import (
     ModelDocument,
     PackageManager,
     PluginManager,
-    ValidateDuplicateFacts,
+    ValidateDuplicateFactsConst,
     Version,
     ViewFileConcepts,
     ViewFileDTS,
@@ -330,11 +330,11 @@ def parseArgs(args: list[str]) -> tuple[RuntimeOptions, dict[str, Any]]:
     validationGroup.add_option(
         "--validateDuplicateFacts",
         "--validateduplicatefacts",
-        choices=[a.value for a in ValidateDuplicateFacts.DUPLICATE_TYPE_ARG_MAP],
+        choices=[a.value for a in ValidateDuplicateFactsConst.DUPLICATE_TYPE_ARG_MAP],
         dest="validateDuplicateFacts",
         help=_(
             "Select which types of duplicates should trigger warnings "
-            f"({' ,'.join([f'`{a.value}`' for a in ValidateDuplicateFacts.DUPLICATE_TYPE_ARG_MAP])})"
+            f"({' ,'.join([f'`{a.value}`' for a in ValidateDuplicateFactsConst.DUPLICATE_TYPE_ARG_MAP])})"
 
             )
         )
@@ -730,11 +730,11 @@ def parseArgs(args: list[str]) -> tuple[RuntimeOptions, dict[str, Any]]:
     outputFileGroup.add_option(
         "--deduplicateFacts",
         "--deduplicatefacts",
-        choices=[a.value for a in ValidateDuplicateFacts.DeduplicationType],
+        choices=[a.value for a in ValidateDuplicateFactsConst.DeduplicationType],
         dest="deduplicateFacts",
         help=_(
             "When using `--saveDeduplicatedInstance` to save a deduplicated instance, check for duplicates of this type"
-            f" ({' ,'.join([f'`{a.value}`' for a in ValidateDuplicateFacts.DeduplicationType])})."
+            f" ({' ,'.join([f'`{a.value}`' for a in ValidateDuplicateFactsConst.DeduplicationType])})."
             )
         )
     outputFileGroup.add_option(
@@ -1811,7 +1811,7 @@ class CntlrCmdLine(Cntlr.Cntlr):
             self.modelManager.baseTaxonomyValidationMode = ValidateBaseTaxonomiesMode.fromName(options.baseTaxonomyValidationMode)
         self.modelManager.validateXmlOim = options.validateXmlOim
         if options.validateDuplicateFacts:
-            duplicateTypeArg = ValidateDuplicateFacts.DuplicateTypeArg(options.validateDuplicateFacts)
+            duplicateTypeArg = ValidateDuplicateFactsConst.DuplicateTypeArg(options.validateDuplicateFacts)
             duplicateType = duplicateTypeArg.duplicateType()
             self.modelManager.validateDuplicateFacts = duplicateType
         self.modelManager.validateAllFilesAsReportPackages = options.reportPackage
@@ -2231,9 +2231,9 @@ class CntlrCmdLine(Cntlr.Cntlr):
                 if success:
                     if options.saveDeduplicatedInstance:
                         if options.deduplicateFacts:
-                            deduplicateFactsArg = ValidateDuplicateFacts.DeduplicationType(options.deduplicateFacts)
+                            deduplicateFactsArg = ValidateDuplicateFactsConst.DeduplicationType(options.deduplicateFacts)
                         else:
-                            deduplicateFactsArg = ValidateDuplicateFacts.DeduplicationType.COMPLETE
+                            deduplicateFactsArg = ValidateDuplicateFactsConst.DeduplicationType.COMPLETE
                         if modelXbrl.modelDocument.type != ModelDocument.Type.INSTANCE:
                             self.addToLog(_("Provided file must be a traditional XBRL instance document to save a deduplicated instance."),
                                           messageCode="error", file=modelXbrl.modelDocument.uri)
@@ -2241,6 +2241,7 @@ class CntlrCmdLine(Cntlr.Cntlr):
                             # Deduplication modifies the underlying lxml tree and leaves the model in an undefined state.
                             # Anything depending on the ModelXbrl that runs after this may encounter unexpected behavior,
                             # so we'll run it as a final step in the CLI controller flow.
+                            from arelle import ValidateDuplicateFacts
                             ValidateDuplicateFacts.saveDeduplicatedInstance(
                                 modelXbrl,
                                 deduplicateFactsArg,
